% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmap_metapone.r
\name{Hmap_metapone}
\alias{Hmap_metapone}
\title{Hmap_metapone
Perform a summary of significant features from metapone pathway tables and generate a heatmap for p-value less than 0.05}
\usage{
Hmap_metapone(path_heatmap, pathway_focus)
}
\arguments{
\item{path_heatmap}{A string representing the path of the input files.}

\item{pathway_focus}{A string representing the pathway focus. For example, V1_VITD_PC. Your final output will be: pathway_focus, "_metapone_overlap", ".csv"}
}
\value{
A data.frame with the metabolite analysis results.
}
\description{
Hmap_metapone
Perform a summary of significant features from metapone pathway tables and generate a heatmap for p-value less than 0.05
}
