% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_reg_2omics_fdr.R
\name{multi_reg_2omics_fdr}
\alias{multi_reg_2omics_fdr}
\title{Multi-Regression Analysis for 2-Omics Data with FDR Correction}
\usage{
multi_reg_2omics_fdr(
  data_comp,
  sig_feature_list,
  outcome_start_column,
  outcome_end_column,
  confounders,
  suffix
)
}
\arguments{
\item{data_comp}{Data frame containing the data for the analysis.}

\item{sig_feature_list}{List of significant features for analysis from the 1st omics results. Make sure the list is a vector}

\item{outcome_start_column}{Integer indicating the starting column number for outcome variables, which is the 2nd omics.}

\item{outcome_end_column}{Integer indicating the ending column number for outcome variables, which is the 2nd omics.}

\item{confounders}{Character vector of confounder variable names. For example, if your interested confounders are "Age_at_enrollment", "Gender", "BMI_on_enrollment", outcome=c("Age_at_enrollment", "Gender", "BMI_on_enrollment"). Note: no space should be in the column names, not even with 'Age at enrollment', which would cause errors.}

\item{suffix}{Character string to be added as a suffix to the output file name.}
}
\value{
A list of data frames containing the results of the multi-regression analysis for each significant feature, saved as CSV files in the working directory.
}
\description{
Multi-Regression Analysis for 2-Omics Data with FDR Correction
}
