% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_lm.R
\name{limma_lm}
\alias{limma_lm}
\title{limma_lm}
\usage{
limma_lm(
  DATA_comp,
  DATA_met = NULL,
  output_name,
  met_start_colnum,
  met_end_colnum,
  outcome,
  confounders
)
}
\arguments{
\item{DATA_comp}{The merged, complete data with demographic data and feature table data path or already-read data. The data can be csv or txt.}

\item{DATA_met}{The cleaned feature table data ready for MWAS analysis. It should be data path or already-read data. The data can be csv or txt.}

\item{output_name}{The name of output file using this function. For example, "Met_meantri_log2_HILIC22"}

\item{met_start_colnum}{The number of the column where metabolites are starting. If metabolite data is starting from column 10, the met_start_colnum=10}

\item{met_end_colnum}{The number of the column where metabolites are ending If metabolite data is ending at column 13414, the met_end_colnum=13414}

\item{outcome}{The outcome of your interest. For example, if your interested outcome is T2D (type 2 diabetes) and the column name is T2D, outcome="T2D"}

\item{confounders}{The confounders of your interest. For example, if your interested confounders are "Age_at_enrollment", "Gender", "BMI_on_enrollment", outcome=c("Age_at_enrollment", "Gender", "BMI_on_enrollment"). Note: no space should be in the column names, not even with 'Age at enrollment', which would cause errors.}
}
\value{
The multiple regression results with FDR correction
}
\description{
Perform Metabolome-wide Association Study (MWAS) analysis using multiple linear regression with FDR correction with the limma algorithm. Other omics can also be used, such as metals from ICP-MS and 16s rRNA microbiome data
}
\examples{
data(stanford_clinical_metabolome)
limma_lm(DATA_comp=stanford_clinical_metabolome, DATA_met=NULL, output_name="stanford_clinical_metabolome", met_start_colnum=58, met_end_colnum=781, outcome="CL4.x", confounders=c("A1C", "GLU"))
}
